+++ internal/app/api/wwapid/wwapid.go
@@ -6,7 +6,6 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"net"
 	"os"
@@ -60,7 +59,7 @@ func main() {
 
 		// Load the CA cert.
 		var cacert []byte
-		cacert, err = ioutil.ReadFile(config.TlsConfig.CaCert)
+		cacert, err = os.ReadFile(config.TlsConfig.CaCert)
 		if err != nil {
 			log.Fatalf("Failed to load cacert. err: %s\n", err)
 		}