+++ internal/pkg/kernel/kernel.go
@@ -4,7 +4,6 @@ import (
 	"compress/gzip"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path"
 	"path/filepath"
@@ -51,7 +50,7 @@ func GetKernelVersion(kernelName string) string {
 		wwlog.Error("Kernel Name is not defined")
 		return ""
 	}
-	kernelVersion, err := ioutil.ReadFile(KernelVersionFile(kernelName))
+	kernelVersion, err := os.ReadFile(KernelVersionFile(kernelName))
 	if err != nil {
 		return ""
 	}
@@ -96,7 +95,7 @@ func ListKernels() ([]string, error) {
 
 	wwlog.Debug("Searching for Kernel image directories: %s", KernelImageTopDir())
 
-	kernels, err := ioutil.ReadDir(KernelImageTopDir())
+	kernels, err := os.ReadDir(KernelImageTopDir())
 	if err != nil {
 		return ret, err
 	}
@@ -198,7 +197,7 @@ func Build(kernelVersion, kernelName, root string) (string, error) {
 			driversDestination,
 			[]string{
 				"." + kernelDriversRelative,
-				"./lib/firmware" },
+				"./lib/firmware"},
 			[]string{},
 			// ignore cross-device files
 			true,